use separate composable for button states
